+++ packages/apps/autorest/src/actions/index.ts
@@ -1,0 +1,2 @@
+export * from "./check-autorest-update";
+export * from "./clear-temp-data";
